requirements.txt
fastapi
uvicorn
pymongo
pandas
python-jose
bcrypt
bleach
python-magic
email-validator
python-multipart
